+++ tests/congruence_tests/test_collections.py
@@ -61,9 +61,11 @@ def test_collection_exists():
 
     vector_params = models.VectorParams(size=2, distance=models.Distance.COSINE)
 
-    remote_client.delete_collection(COLLECTION_NAME)
+    if remote_client.collection_exists(COLLECTION_NAME):
+        remote_client.delete_collection(COLLECTION_NAME)
     remote_client.create_collection(COLLECTION_NAME, vectors_config=vector_params)
-    local_client.delete_collection(COLLECTION_NAME)
+    if local_client.collection_exists(COLLECTION_NAME):
+        local_client.delete_collection(COLLECTION_NAME)
     local_client.create_collection(COLLECTION_NAME, vectors_config=vector_params)
 
     assert remote_client.collection_exists(COLLECTION_NAME)
@@ -79,20 +81,24 @@ def test_init_from():
     points = generate_fixtures(vectors_sizes=vector_size)
     vector_params = models.VectorParams(size=vector_size, distance=models.Distance.COSINE)
 
-    remote_client.delete_collection(collection_name=COLLECTION_NAME)
+    if remote_client.collection_exists(COLLECTION_NAME):
+        remote_client.delete_collection(collection_name=COLLECTION_NAME)
     remote_client.create_collection(collection_name=COLLECTION_NAME, vectors_config=vector_params)
-    local_client.delete_collection(collection_name=COLLECTION_NAME)
+    if local_client.collection_exists(COLLECTION_NAME):
+        local_client.delete_collection(collection_name=COLLECTION_NAME)
     local_client.create_collection(collection_name=COLLECTION_NAME, vectors_config=vector_params)
     remote_client.upload_points(COLLECTION_NAME, points, wait=True)
     local_client.upload_points(COLLECTION_NAME, points)
     compare_collections(remote_client, local_client, len(points), collection_name=COLLECTION_NAME)
 
     new_collection_name = COLLECTION_NAME + "_new"
-    remote_client.delete_collection(new_collection_name)
+    if remote_client.collection_exists(new_collection_name):
+        remote_client.delete_collection(new_collection_name)
     remote_client.create_collection(
         new_collection_name, vectors_config=vector_params, init_from=COLLECTION_NAME
     )
-    local_client.delete_collection(new_collection_name)
+    if local_client.collection_exists(new_collection_name):
+        local_client.delete_collection(new_collection_name)
     local_client.create_collection(
         new_collection_name, vectors_config=vector_params, init_from=COLLECTION_NAME
     )
@@ -106,13 +112,15 @@ def test_init_from():
         collection_name=new_collection_name,
     )
 
-    remote_client.delete_collection(new_collection_name)
+    if remote_client.collection_exists(new_collection_name):
+        remote_client.delete_collection(new_collection_name)
     remote_client.create_collection(
         new_collection_name,
         vectors_config=vector_params,
         init_from=models.InitFrom(collection=COLLECTION_NAME),
     )
-    local_client.delete_collection(new_collection_name)
+    if local_client.collection_exists(new_collection_name):
+        local_client.delete_collection(new_collection_name)
     local_client.create_collection(
         new_collection_name,
         vectors_config=vector_params,
