+++ tests/congruence_tests/test_sparse_search.py
@@ -297,11 +297,13 @@ def test_query_with_nan():
     )
     named_sparse_vector.vector.values[0] = np.nan
 
-    local_client.delete_collection(COLLECTION_NAME)
+    if local_client.collection_exists(COLLECTION_NAME):
+        local_client.delete_collection(COLLECTION_NAME)
     local_client.create_collection(
         COLLECTION_NAME, vectors_config={}, sparse_vectors_config=sparse_vectors_config
     )
-    remote_client.delete_collection(COLLECTION_NAME)
+    if remote_client.collection_exists(COLLECTION_NAME):
+        remote_client.delete_collection(COLLECTION_NAME)
     remote_client.create_collection(
         COLLECTION_NAME, vectors_config={}, sparse_vectors_config=sparse_vectors_config
     )
