+++ tests/congruence_tests/test_updates.py
@@ -158,9 +158,11 @@ def test_upload_collection_float_list():
 
     vectors = np.random.randn(UPLOAD_NUM_VECTORS, vectors_dim).tolist()
     vectors_config = models.VectorParams(size=vectors_dim, distance=models.Distance.EUCLID)
-    local_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
+    if local_client.collection_exists(COLLECTION_NAME):
+        local_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
     local_client.create_collection(COLLECTION_NAME, vectors_config=vectors_config, timeout=TIMEOUT)
-    remote_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
+    if remote_client.collection_exists(COLLECTION_NAME):
+        remote_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
     remote_client.create_collection(
         COLLECTION_NAME, vectors_config=vectors_config, timeout=TIMEOUT
     )
@@ -193,13 +195,15 @@ def test_upload_collection_np_array_2d():
     vectors = np.random.randn(UPLOAD_NUM_VECTORS, vectors_dim)
     ids = list(range(len(vectors)))
     vectors_config = models.VectorParams(size=vectors_dim, distance=models.Distance.EUCLID)
-    local_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
+    if local_client.collection_exists(COLLECTION_NAME):
+        local_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
     local_client.create_collection(
         COLLECTION_NAME,
         vectors_config=vectors_config,
         timeout=TIMEOUT,
     )
-    remote_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
+    if remote_client.collection_exists(COLLECTION_NAME):
+        remote_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
     remote_client.create_collection(
         COLLECTION_NAME,
         vectors_config=vectors_config,
@@ -223,13 +227,15 @@ def test_upload_collection_list_np_arrays():
     vectors = [np.array(vector) for vector in vectors]
     vectors_config = models.VectorParams(size=vectors_dim, distance=models.Distance.EUCLID)
     ids = list(range(len(vectors)))
-    local_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
+    if local_client.collection_exists(COLLECTION_NAME):
+        local_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
     local_client.create_collection(
         COLLECTION_NAME,
         vectors_config=vectors_config,
         timeout=TIMEOUT,
     )
-    remote_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
+    if remote_client.collection_exists(COLLECTION_NAME):
+        remote_client.delete_collection(COLLECTION_NAME, timeout=TIMEOUT)
     remote_client.create_collection(
         COLLECTION_NAME,
         vectors_config=vectors_config,
@@ -271,13 +277,15 @@ def test_upload_wrong_vectors():
         "text": models.VectorParams(size=vector_size, distance=models.Distance.COSINE)
     }
     sparse_vectors_config = {"text-sparse": models.SparseVectorParams()}
-    local_client.delete_collection(collection_name=wrong_vectors_collection)
+    if local_client.collection_exists(collection_name=wrong_vectors_collection):
+        local_client.delete_collection(collection_name=wrong_vectors_collection)
     local_client.create_collection(
         collection_name=wrong_vectors_collection,
         vectors_config=vectors_config,
         sparse_vectors_config=sparse_vectors_config,
     )
-    remote_client.delete_collection(collection_name=wrong_vectors_collection)
+    if remote_client.collection_exists(collection_name=wrong_vectors_collection):
+        remote_client.delete_collection(collection_name=wrong_vectors_collection)
     remote_client.create_collection(
         collection_name=wrong_vectors_collection,
         vectors_config=vectors_config,
