+++ tests/test_async_qdrant_client.py
@@ -111,7 +111,8 @@ async def test_async_qdrant_client(prefer_grpc):
         qdrant_client.http.exceptions.UnexpectedResponse,
         grpc.aio._call.AioRpcError,
     ):
-        await client.delete_collection(COLLECTION_NAME)
+        if await client.collection_exists(COLLECTION_NAME):
+            await client.delete_collection(COLLECTION_NAME)
         await client.create_collection(**collection_params)
 
     await client.get_collection(COLLECTION_NAME)
@@ -336,7 +337,8 @@ async def test_async_qdrant_client(prefer_grpc):
     assert (await client.retrieve(COLLECTION_NAME, ids=[1]))[0].payload == {}
 
     # region teardown
-    await client.delete_collection(COLLECTION_NAME)
+    if await client.collection_exists(COLLECTION_NAME):
+        await client.delete_collection(COLLECTION_NAME)
     collections = await client.get_collections()
 
     assert all(collection.name != COLLECTION_NAME for collection in collections.collections)
@@ -354,7 +356,8 @@ async def test_async_qdrant_client_local():
         collection_name=COLLECTION_NAME,
         vectors_config=models.VectorParams(size=10, distance=models.Distance.EUCLID),
     )
-    await client.delete_collection(COLLECTION_NAME)
+    if await client.collection_exists(COLLECTION_NAME):
+        await client.delete_collection(COLLECTION_NAME)
     await client.create_collection(**collection_params)
 
     await client.get_collection(COLLECTION_NAME)
@@ -535,7 +538,8 @@ async def test_async_qdrant_client_local():
     assert (await client.retrieve(COLLECTION_NAME, ids=[1]))[0].payload == {}
 
     # region teardown
-    await client.delete_collection(COLLECTION_NAME)
+    if await client.collection_exists(COLLECTION_NAME):
+        await client.delete_collection(COLLECTION_NAME)
     collections = await client.get_collections()
 
     assert all(collection.name != COLLECTION_NAME for collection in collections.collections)
