+++ tests/test_migrate.py
@@ -383,7 +383,8 @@ def test_vector_params(
         ),
     }
 
-    local_client.delete_collection(collection_name=collection_name)
+    if local_client.collection_exists(collection_name):
+        local_client.delete_collection(collection_name=collection_name)
     local_client.create_collection(collection_name=collection_name, vectors_config=vectors_config)
 
     local_client.migrate(second_local_client)
