+++ tests/test_qdrant_client.py
@@ -199,8 +199,8 @@ def test_records_upload(prefer_grpc, parallel):
     )
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -235,7 +235,8 @@ def test_records_upload(prefer_grpc, parallel):
 
     records = (Record(id=idx, vector=np.random.rand(DIM).tolist()) for idx in range(NUM_VECTORS))
 
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -265,7 +266,8 @@ def test_point_upload(prefer_grpc, parallel):
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
 
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -297,8 +299,8 @@ def test_point_upload(prefer_grpc, parallel):
 
     assert result_count.count < 900
     assert result_count.count > 100
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -325,7 +327,8 @@ def test_upload_collection(prefer_grpc, parallel):
     batch_size = 2
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
 
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=size, distance=Distance.DOT),
@@ -350,8 +353,8 @@ def test_upload_collection(prefer_grpc, parallel):
     )
 
     assert client.get_collection(collection_name=COLLECTION_NAME).points_count == 5
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=size, distance=Distance.DOT),
@@ -387,8 +390,8 @@ def test_multiple_vectors(prefer_grpc):
     ]
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={
@@ -447,8 +450,8 @@ def test_qdrant_client_integration(prefer_grpc, numpy_upload, local_mode):
         client = QdrantClient(location=":memory:", prefer_grpc=prefer_grpc)
     else:
         client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1040,8 +1043,8 @@ def test_qdrant_client_integration(prefer_grpc, numpy_upload, local_mode):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_qdrant_client_integration_update_collection(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={
@@ -1101,8 +1104,8 @@ def test_qdrant_client_integration_update_collection(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_points_crud(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1151,8 +1154,8 @@ def test_points_crud(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_quantization_config(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1212,7 +1215,8 @@ def test_custom_sharding(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
 
     def init_collection():
-        client.delete_collection(collection_name=COLLECTION_NAME)
+        if client.collection_exists(COLLECTION_NAME):
+            client.delete_collection(collection_name=COLLECTION_NAME)
         client.create_collection(
             collection_name=COLLECTION_NAME,
             vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1387,8 +1391,8 @@ def test_sparse_vectors(prefer_grpc):
         pytest.skip("Sparse vectors are supported since v1.7.0")
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={},
@@ -1467,8 +1471,8 @@ def test_sparse_vectors_batch(prefer_grpc):
         pytest.skip("Sparse vectors are supported since v1.7.0")
 
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={},
@@ -1550,8 +1554,8 @@ def test_sparse_vectors_batch(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_vector_update(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1610,8 +1614,8 @@ def test_vector_update(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_conditional_payload_update(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1650,8 +1654,8 @@ def test_conditional_payload_update(prefer_grpc):
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_conditional_payload_update_2(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if not client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1731,8 +1735,8 @@ def test_insert_float():
 
 def test_locks():
     client = QdrantClient(timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config=VectorParams(size=DIM, distance=Distance.DOT),
@@ -1778,8 +1782,8 @@ def test_locks():
 @pytest.mark.parametrize("prefer_grpc", [False, True])
 def test_empty_vector(prefer_grpc):
     client = QdrantClient(prefer_grpc=prefer_grpc, timeout=TIMEOUT)
-
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=TIMEOUT)
     client.create_collection(
         collection_name=COLLECTION_NAME,
         vectors_config={},
@@ -1848,13 +1852,15 @@ def test_client_close():
 
     # region http
     client_http = QdrantClient(timeout=TIMEOUT)
-    client_http.delete_collection("test")
+    if client_http.collection_exists("test"):
+        client_http.delete_collection("test")
     client_http.create_collection(
         "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
     )
     client_http.close()
     with pytest.raises(qdrant_exceptions.ResponseHandlingException):
-        client_http.delete_collection("test")
+        if client_http.collection_exists("test"):
+            client_http.delete_collection("test")
         client_http.create_collection(
             "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
         )
@@ -1862,7 +1868,8 @@ def test_client_close():
 
     # region grpc
     client_grpc = QdrantClient(prefer_grpc=True, timeout=TIMEOUT)
-    client_grpc.delete_collection("test")
+    if client_grpc.collection_exists("test"):
+        client_grpc.delete_collection("test")
     client_grpc.create_collection(
         "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
     )
@@ -1904,7 +1911,8 @@ def test_client_close():
 
     # region local
     local_client_in_mem = QdrantClient(":memory:")
-    local_client_in_mem.delete_collection("test")
+    if local_client_in_mem.collection_exists("test"):
+        local_client_in_mem.delete_collection("test")
     local_client_in_mem.create_collection(
         "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
     )
@@ -1922,20 +1930,23 @@ def test_client_close():
         )
 
     with pytest.raises(RuntimeError):
-        local_client_in_mem.delete_collection("test")
+        if local_client_in_mem.collection_exists("test"):
+            local_client_in_mem.delete_collection("test")
 
     with tempfile.TemporaryDirectory() as tmpdir:
         path = tmpdir + "/test.db"
 
         local_client_persist_1 = QdrantClient(path=path)
-        local_client_persist_1.delete_collection("test")
+        if local_client_persist_1.collection_exists("test"):
+            local_client_persist_1.delete_collection("test")
         local_client_persist_1.create_collection(
             "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
         )
         local_client_persist_1.close()
 
         local_client_persist_2 = QdrantClient(path=path)
-        local_client_persist_2.delete_collection("test")
+        if local_client_persist_2.collection_exists("test"):
+            local_client_persist_2.delete_collection("test")
         local_client_persist_2.create_collection(
             "test", vectors_config=VectorParams(size=100, distance=Distance.COSINE)
         )
@@ -1956,10 +1967,12 @@ def test_timeout_propagation():
         # timeout is Optional[int]
         # if we set it to 0 - recreate_collection raises operation is in progress instead of timed out
         client.http.client._client._timeout = Timeout(0.01)
-        client.delete_collection(collection_name=COLLECTION_NAME)
+        if client.collection_exists(COLLECTION_NAME):
+            client.delete_collection(collection_name=COLLECTION_NAME)
         client.create_collection(collection_name=COLLECTION_NAME, vectors_config=vectors_config)
     time.sleep(0.5)
-    client.delete_collection(collection_name=COLLECTION_NAME, timeout=10)
+    if client.collection_exists(COLLECTION_NAME):
+        client.delete_collection(collection_name=COLLECTION_NAME, timeout=10)
     client.create_collection(
         collection_name=COLLECTION_NAME, vectors_config=vectors_config, timeout=10
     )
